+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Modal: Prop 'size' adds class to div.modal-window instead of style attribute
 
+### Fixed
+- DatePicker: Fixed date selection from previous month
 
 ## [0.40.0] - 2020-02-09